+++ bin/faucet/src/main.rs
@@ -1,28 +1,30 @@
 mod faucet;
-mod rpc_client;
 mod server;
 mod types;
 
 mod network;
 #[cfg(test)]
-mod stub_rpc_api;
+mod testing;
 
-use std::{num::NonZeroUsize, path::PathBuf, time::Duration};
+use std::{num::NonZeroUsize, path::PathBuf, sync::Arc, time::Duration};
 
 use anyhow::Context;
 use clap::{Parser, Subcommand};
 use faucet::Faucet;
-use miden_lib::{AuthScheme, account::faucets::create_basic_fungible_faucet};
-use miden_node_utils::{crypto::get_rpo_random_coin, logging::OpenTelemetry, version::LongVersion};
-use miden_objects::{
+use miden_client::{
     Felt,
-    account::{AccountFile, AccountStorageMode, AuthSecretKey},
+    account::{
+        AccountBuilder, AccountFile, AccountStorageMode, AccountType,
+        component::{BasicFungibleFaucet, RpoFalcon512},
+    },
     asset::TokenSymbol,
-    crypto::dsa::rpo_falcon512::SecretKey,
+    auth::AuthSecretKey,
+    crypto::SecretKey,
+    store::sqlite_store::SqliteStore,
 };
+use miden_node_utils::{crypto::get_rpo_random_coin, logging::OpenTelemetry, version::LongVersion};
 use rand::{Rng, SeedableRng};
 use rand_chacha::ChaCha20Rng;
-use rpc_client::RpcClient;
 use server::Server;
 use tokio::sync::mpsc;
 use types::AssetOptions;
@@ -53,6 +55,7 @@ const ENV_POW_BASELINE: &str = "MIDEN_FAUCET_POW_BASELINE";
 const ENV_API_KEYS: &str = "MIDEN_FAUCET_API_KEYS";
 const ENV_ENABLE_OTEL: &str = "MIDEN_FAUCET_ENABLE_OTEL";
 const ENV_NETWORK: &str = "MIDEN_FAUCET_NETWORK";
+const ENV_STORE: &str = "MIDEN_FAUCET_STORE";
 
 // COMMANDS
 // ================================================================================================
@@ -134,6 +137,10 @@ pub enum Command {
         /// OpenTelemetry documentation. See our operator manual for further details.
         #[arg(long = "enable-otel", value_name = "BOOL", default_value_t = false, env = ENV_ENABLE_OTEL)]
         open_telemetry: bool,
+
+        /// Path to the `SQLite` store.
+        #[arg(long = "store", value_name = "FILE", default_value = "faucet_client_store.sqlite3", env = ENV_STORE)]
+        store_path: PathBuf,
     },
 
     /// Create a new public faucet account and save to the specified file.
@@ -205,24 +212,27 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
             pow_baseline,
             api_keys,
             open_telemetry: _,
+            store_path,
         } => {
-            let mut rpc_client = RpcClient::connect_lazy(&node_url, timeout.as_millis() as u64)
-                .context("failed to create RPC client")?;
             let account_file = AccountFile::read(&faucet_account_path).context(format!(
                 "failed to load faucet account from file ({})",
                 faucet_account_path.display()
             ))?;
 
             let faucet = Faucet::load(
+                store_path.clone(),
                 network.to_network_id()?,
                 account_file,
-                &mut rpc_client,
+                &node_url,
+                timeout,
                 remote_tx_prover_url,
             )
             .await?;
+            let store =
+                Arc::new(SqliteStore::new(store_path).await.context("failed to create store")?);
 
             // Maximum of 1000 requests in-queue at once. Overflow is rejected for faster feedback.
-            let (tx_requests, rx_requests) = mpsc::channel(REQUESTS_QUEUE_SIZE);
+            let (tx_mint_requests, rx_mint_requests) = mpsc::channel(REQUESTS_QUEUE_SIZE);
 
             let api_keys = api_keys
                 .iter()
@@ -240,16 +250,17 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
             let server = Server::new(
                 faucet.faucet_id(),
                 asset_options,
-                tx_requests,
+                tx_mint_requests,
                 pow_secret.unwrap_or_default().as_str(),
                 pow_config,
                 &api_keys,
+                store,
             );
 
             // Use select to concurrently:
             // - Run and wait for the faucet (on current thread)
             // - Run and wait for server (in a spawned task)
-            let faucet_future = faucet.run(rpc_client, rx_requests);
+            let faucet_future = faucet.run(rx_mint_requests);
             let server_future = async {
                 let server_handle =
                     tokio::spawn(
@@ -266,7 +277,7 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
                 faucet_result = faucet_future => {
                     // Faucet completed, return its result
                     faucet_result.context("faucet failed")
-                }
+                },
             }?;
         },
 
@@ -285,17 +296,19 @@ async fn run_faucet_command(cli: Cli) -> anyhow::Result<()> {
 
             let secret = SecretKey::with_rng(&mut get_rpo_random_coin(&mut rng));
 
-            let (account, account_seed) = create_basic_fungible_faucet(
-                rng.random(),
-                TokenSymbol::try_from(token_symbol.as_str())
-                    .context("failed to parse token symbol")?,
-                decimals,
-                Felt::try_from(max_supply)
-                    .expect("max supply value is greater than or equal to the field modulus"),
-                AccountStorageMode::Public,
-                AuthScheme::RpoFalcon512 { pub_key: secret.public_key() },
-            )
-            .context("failed to create basic fungible faucet account")?;
+            let symbol = TokenSymbol::try_from(token_symbol.as_str())
+                .context("failed to parse token symbol")?;
+            let max_supply = Felt::try_from(max_supply)
+                .map_err(anyhow::Error::msg)
+                .context("max supply value is greater than or equal to the field modulus")?;
+
+            let (account, account_seed) = AccountBuilder::new(rng.random())
+                .account_type(AccountType::FungibleFaucet)
+                .storage_mode(AccountStorageMode::Public)
+                .with_component(BasicFungibleFaucet::new(symbol, decimals, max_supply)?)
+                .with_auth_component(RpoFalcon512::new(secret.public_key()))
+                .build()
+                .context("failed to create basic fungible faucet account")?;
 
             let account_data = AccountFile::new(
                 account,
@@ -346,6 +359,7 @@ mod test {
     use std::{
         env::temp_dir,
         num::NonZeroUsize,
+        path::PathBuf,
         process::Stdio,
         str::FromStr,
         time::{Duration, Instant},
@@ -357,7 +371,7 @@ mod test {
     use tokio::{io::AsyncBufReadExt, time::sleep};
     use url::Url;
 
-    use crate::{Cli, FaucetNetwork, run_faucet_command, stub_rpc_api::serve_stub};
+    use crate::{Cli, FaucetNetwork, run_faucet_command, testing::stub_rpc_api::serve_stub};
 
     /// This test starts a stub node, a faucet connected to the stub node, and a chromedriver
     /// to test the faucet website. It then loads the website and checks that all the requests
@@ -407,7 +421,7 @@ mod test {
 
         // Fill in the account address
         client
-            .find(fantoccini::Locator::Css("#recipientAddress"))
+            .find(fantoccini::Locator::Css("#recipient-address"))
             .await
             .unwrap()
             .send_keys("mtst1qrvhealccdyj7gqqqrlxl4n4f53uxwaw")
@@ -416,14 +430,14 @@ mod test {
 
         // Select the first asset amount option
         client
-            .find(fantoccini::Locator::Css("#tokenAmount"))
+            .find(fantoccini::Locator::Css("#token-amount"))
             .await
             .unwrap()
             .click()
             .await
             .unwrap();
         client
-            .find(fantoccini::Locator::Css("#tokenAmount option"))
+            .find(fantoccini::Locator::Css("#token-amount option"))
             .await
             .unwrap()
             .click()
@@ -432,7 +446,7 @@ mod test {
 
         // Click the public note button
         client
-            .find(fantoccini::Locator::Css("#sendPublicButton"))
+            .find(fantoccini::Locator::Css("#send-public-button"))
             .await
             .unwrap()
             .click()
@@ -451,7 +465,7 @@ mod test {
                 .as_array()
                 .unwrap()
                 .clone();
-            if events.iter().any(|event| event["type"] == "note") {
+            if events.iter().any(|event| event["type"] == "minted") {
                 captured_events = events;
                 break;
             }
@@ -460,11 +474,10 @@ mod test {
 
         // Verify the received events
         assert!(!captured_events.is_empty(), "Took too long to capture any events");
-        let note_event = captured_events.iter().find(|event| event["type"] == "note").unwrap();
+        let note_event = captured_events.iter().find(|event| event["type"] == "minted").unwrap();
         let note_data: serde_json::Value =
             serde_json::from_str(note_event["data"].as_str().unwrap()).unwrap();
         assert!(note_data["note_id"].is_string());
-        assert!(note_data["account_id"].is_string());
         assert!(note_data["transaction_id"].is_string());
         assert!(note_data["explorer_url"].is_string());
 
@@ -522,6 +535,7 @@ mod test {
                         faucet_account_path: faucet_account_path.clone(),
                         remote_tx_prover_url: None,
                         open_telemetry: false,
+                        store_path: PathBuf::from("faucet_client_store.sqlite3"),
                     },
                 })
                 .await