+++ bin/faucet/src/server/get_pow.rs
@@ -4,7 +4,7 @@ use axum::{
     response::IntoResponse,
 };
 use http::StatusCode;
-use miden_objects::{AccountIdError, account::AccountId};
+use miden_client::account::{AccountId, AccountIdError};
 use serde::Deserialize;
 
 use crate::server::{ApiKey, pow::PoW};
@@ -15,7 +15,7 @@ use crate::server::{ApiKey, pow::PoW};
 pub async fn get_pow(
     State(pow): State<PoW>,
     Query(params): Query<RawPowRequest>,
-) -> Result<impl IntoResponse, MintRequestError> {
+) -> Result<impl IntoResponse, PowRequestError> {
     let request = params.validate()?;
     let challenge = pow.build_challenge(request);
     Ok(Json(challenge))
@@ -38,20 +38,20 @@ pub struct PowRequest {
 }
 
 impl RawPowRequest {
-    pub fn validate(self) -> Result<PowRequest, MintRequestError> {
+    pub fn validate(self) -> Result<PowRequest, PowRequestError> {
         let account_id = if self.account_id.starts_with("0x") {
             AccountId::from_hex(&self.account_id)
         } else {
             AccountId::from_bech32(&self.account_id).map(|(_, account_id)| account_id)
         }
-        .map_err(MintRequestError::InvalidAccount)?;
+        .map_err(PowRequestError::InvalidAccount)?;
 
         let api_key = self
             .api_key
             .as_deref()
             .map(ApiKey::decode)
             .transpose()
-            .map_err(|_| MintRequestError::InvalidApiKey(self.api_key.unwrap_or_default()))?
+            .map_err(|_| PowRequestError::InvalidApiKey(self.api_key.unwrap_or_default()))?
             .unwrap_or_default();
 
         Ok(PowRequest { account_id, api_key })
@@ -59,14 +59,14 @@ impl RawPowRequest {
 }
 
 #[derive(Debug, thiserror::Error)]
-pub enum MintRequestError {
+pub enum PowRequestError {
     #[error("account address failed to parse")]
     InvalidAccount(#[source] AccountIdError),
     #[error("API key failed to parse")]
     InvalidApiKey(String),
 }
 
-impl MintRequestError {
+impl PowRequestError {
     /// Take care to not expose internal errors here.
     fn user_facing_error(&self) -> String {
         match self {
@@ -76,7 +76,7 @@ impl MintRequestError {
     }
 }
 
-impl IntoResponse for MintRequestError {
+impl IntoResponse for PowRequestError {
     fn into_response(self) -> axum::response::Response {
         (StatusCode::BAD_REQUEST, self.user_facing_error()).into_response()
     }