+++ CHANGELOG.md
@@ -4,8 +4,11 @@
 
 ### Changes
 
-- Redesigned the home frontend (#20).
+- Integrated miden-client ([#11](https://github.com/0xMiden/miden-faucet/pull/11)).
+- Added `/get_note` endpoint ([#19](https://github.com/0xMiden/miden-faucet/pull/19)).
+- Redesigned the home frontend ([#20](https://github.com/0xMiden/miden-faucet/pull/20)).
+- Redesigned the tokens request flows ([#25](https://github.com/0xMiden/miden-faucet/pull/25)).
 
 ## 0.10.0 (2025-07-10)
 
-For previous changes see https://github.com/0xMiden/miden-node/blob/main/CHANGELOG.md
+For previous changes see the changelog in https://github.com/0xMiden/miden-node.